[YS-473] refactor: 참여자 내정보 수정 연락받을 이메일 중복 확인 로직 개선
